tentative changes to the function and class definition, need testing in lab to see the most efficient usages for the function/constructor

+++ classes/Rubisco_Chain.m
@@ -10,6 +10,7 @@ classdef Rubisco_Chain < handle
         tag = nan;
         tags uint32 = uint32([]); % tags of rubiscos in chain
         indices uint32 = uint32([]); % indices of rubiscos in chain
+        % Primarily testing two ways - tags using casting initialization, indices using direct initialization in the constructor
         average_vector = []; % average vector of rubiscos in chain
         centroid = []; % centroid of chain
         eigenvalues = []; % eigenvalues of the S tensor
@@ -27,7 +28,7 @@ classdef Rubisco_Chain < handle
                 chain.reg = reg;
                 chain.tags = tags;
                 chain.length = length(tags);
-                chain.indices = indices;
+                chain.indices = uint32(indices);
                 chain.type = type;
                 chain.index = index;
                 chain.tag = tag;
